% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{add_1940_1950}
\alias{add_1940_1950}
\title{internal fucntion to add early years to population data}
\usage{
add_1940_1950(pop2d)
}
\arguments{
\item{pop2d}{2-dimensional population data}
}
\value{
2-dimensional population data with early years
}
\description{
internal fucntion to add early years to population data
}
