% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{calc_incidence_vac_general}
\alias{calc_incidence_vac_general}
\title{calculate the first instance of vaccine coverage}
\usage{
calc_incidence_vac_general(vc3d)
}
\arguments{
\item{vc3d}{vaccination data in array form}
}
\value{
inc_v3d
}
\description{
calculate the first instance of vaccine coverage
}
