% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{calc_pop_moments}
\alias{calc_pop_moments}
\title{calculate population moments}
\usage{
calc_pop_moments(p_prop, t0_vac_, dim_adm, dim_year, dim_age)
}
\arguments{
\item{p_prop}{proportion of population}

\item{t0_vac_}{intial vaccination time in area of interest}

\item{dim_adm}{admins of interest}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}
}
\value{
pop_moments_whole
}
\description{
calculate population moments
}
