% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{calc_pop_moments_agg}
\alias{calc_pop_moments_agg}
\title{calculate the population moments for aggregated data}
\usage{
calc_pop_moments_agg(pop_agg3d, t0_vac_, dim_year, study_years)
}
\arguments{
\item{pop_agg3d}{population aggregated in array}

\item{t0_vac_}{initial vaccination time for area of interest}

\item{dim_year}{years of interest}

\item{study_years}{years of serological studies}
}
\value{
pop_moments_agg
}
\description{
calculate the population moments for aggregated data
}
