% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{calc_t0_vac_africa}
\alias{calc_t0_vac_africa}
\title{get time of first vaccination}
\usage{
calc_t0_vac_africa(vc3d)
}
\arguments{
\item{vc3d}{vaccination data in array form}
}
\value{
t0_vac_africa
}
\description{
get time of first vaccination
}
