% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_functions.R
\name{calculate_bootstrap_Bayes}
\alias{calculate_bootstrap_Bayes}
\title{Calculate Bayes factors and bootstrap for uncertainty}
\usage{
calculate_bootstrap_Bayes(mcmc_out, col = "orchid", plotout = TRUE)
}
\arguments{
\item{mcmc_out}{chains from serology}

\item{col}{colour for histogram, defaults to "orchid"}

\item{plotout}{whether to plot or not- defaults to TRUE}
}
\value{
histogram of model evidence, proportion of iterations
        spent on R0 model and log_Bayes_R0_Foi
}
\description{
Calculate Bayes factors and bootstrap for uncertainty
}
