% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_functions.R
\name{calculate_hpd}
\alias{calculate_hpd}
\title{Calculating the High probability density regions for all parameters}
\usage{
calculate_hpd(mcmc_out, glm_mcmc_out)
}
\arguments{
\item{mcmc_out}{chains from production space serology estimation}

\item{glm_mcmc_out}{chains from glm estimation}
}
\value{
hpd for R0 and Foi models with glm parameters
}
\description{
Calculating the High probability density regions for all parameters
}
