% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{create_pop30_agg_vc30_agg}
\alias{create_pop30_agg_vc30_agg}
\title{get aggregated population and vaccination across observation period}
\usage{
create_pop30_agg_vc30_agg(pop1, vc2d)
}
\arguments{
\item{pop1}{population data}

\item{vc2d}{vaccine coverage data}
}
\value{
pop30_agg, pop_vc_moments and vc30_agg, population and vaccination aggregated over observation period
}
\description{
get aggregated population and vaccination across observation period
}
