% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{create_pop_at_survey}
\alias{create_pop_at_survey}
\title{get the population at the survey times and places}
\usage{
create_pop_at_survey(pop_agg3d, dim_survey, dim_year)
}
\arguments{
\item{pop_agg3d}{aggregated population}

\item{dim_survey}{surveys of interest}

\item{dim_year}{years of interest}
}
\value{
p_at_survey_3d, P_tot_survey_2d
}
\description{
get the population at the survey times and places
}
