% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0_model_functions.R
\name{create_R0_lookup}
\alias{create_R0_lookup}
\title{function to calculate aggregated seroprevalence at survey in R0 model}
\usage{
create_R0_lookup(dim_adm, dat, dim_year, dim_age, p_prop_3d, P_tot_2d,
  inc_v3d, pop_moments_whole, pop_moments_agg, vac_eff_arg = 1,
  granularity = "coarse")
}
\arguments{
\item{dim_adm}{admin 1 units}

\item{dat}{data on environment and occurrence}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{inc_v3d}{incidence of vacciantion}

\item{pop_moments_whole}{popualtion moments}

\item{pop_moments_agg}{aggregated population moments}

\item{vac_eff_arg}{vaccine efficacy}

\item{granularity}{how many intervals in R0 values, defaults to "coarse"}

\item{p_at_survey_3d}{population proportion  as an array}

\item{P_tot_survey_2d}{population total}

\item{vcfac_arg}{vaccine coverage factor, defaults to 1}
}
\value{
file of R0 related to number of infections over the who time period
}
\description{
function to calculate aggregated seroprevalence at survey in R0 model
}
