% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{fit_glm}
\alias{fit_glm}
\title{fit the glm}
\usage{
fit_glm(dat, depi, models)
}
\arguments{
\item{dat}{data with wenvironmental and occurrence information}

\item{depi}{dependant variables}

\item{models}{list of models- only use first one}
}
\value{
coefficients of glm
}
\description{
fit the glm
}
