% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Foi_model_functions.R
\name{foi_calculate_seroprevalence}
\alias{foi_calculate_seroprevalence}
\title{function to calculate aggregated seroprevalence at survey in foi model}
\usage{
foi_calculate_seroprevalence(foi, age_groups, pop, vc)
}
\arguments{
\item{foi}{value of force of infection for survey}

\item{age_groups}{age groups in survey}

\item{pop}{population in year and place of survey}

\item{vc}{vaccine coverage, defaults to 0}
}
\value{
seroprevalence at age groups
}
\description{
function to calculate aggregated seroprevalence at survey in foi model
}
