% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_functions.R
\name{fun_calc_pdetect_multi_both}
\alias{fun_calc_pdetect_multi_both}
\title{calculate the probability of detection for both models}
\usage{
fun_calc_pdetect_multi_both(x, seroout, params, dat, t0_vac_africa,
  dim_year, dim_age, p_prop_3d, P_tot_2d, inc_v3d, pop_moments_whole,
  varsin_nc, vc30_agg, pop30_agg, model_type)
}
\arguments{
\item{x}{glm covariates}

\item{seroout}{processed serology data}

\item{params}{estimated parameters for both serology and GLM from either R0 or Foi model}

\item{dat}{environmental and occurrence data}

\item{t0_vac_africa}{time of first vaccination for admins in Africa endemic zone}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{p_prop_3d}{prpoprtion of population in array form}

\item{P_tot_2d}{total population in 2d form}

\item{pop_moments_whole}{population moments for all admins}

\item{varsin_nc}{covariates in use}

\item{vc30_agg}{vaccine coverage aggregated over time and space for observation period}

\item{pop30_agg}{population aggregated over time and space for observation period}

\item{model_type}{either "R0" or "Foi"}

\item{inc_vc3d}{incidence of vaccination in array form}
}
\value{
the probability of detection
}
\description{
calculate the probability of detection for both models
}
