% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_functions.R
\name{fun_calc_transmission_Africa}
\alias{fun_calc_transmission_Africa}
\title{calculate the probability of detection for both models}
\usage{
fun_calc_transmission_Africa(x, ii, seroout, params, dat, t0_vac_africa,
  dim_year, dim_age, p_prop_3d, P_tot_2d, inc_v3d, pop1, vc2d, varsin_nc,
  polydeg = 5, R0_lookup, model_type)
}
\arguments{
\item{x}{glm covariates}

\item{ii}{indices of glm parameters}

\item{seroout}{processed serology data}

\item{params}{estimated parameters for both serology and GLM from either R0 or Foi model}

\item{dat}{environmental and occurrence data}

\item{t0_vac_africa}{time of first vaccination for admins in Africa endemic zone}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{p_prop_3d}{prpoprtion of population in array form}

\item{P_tot_2d}{total population in 2d form}

\item{pop1}{population data}

\item{vc2d}{vaccination coverages in 2d form}

\item{varsin_nc}{covariates in use}

\item{polydeg}{degree of polynomial, defaults to 5}

\item{R0_lookup}{table of R0 and infection numbers for each admin}

\item{model_type}{either "R0" or "Foi"}

\item{inc_vc3d}{incidence of vaccination in array form}
}
\value{
transmission_whole, transmission intensty across Africa endemic zone
}
\description{
calculate the probability of detection for both models
}
