% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_functions.R
\name{fun_calcPred}
\alias{fun_calcPred}
\title{calculate predicitions from glm}
\usage{
fun_calcPred(coefs, newdata, type = "response", varsin = NA)
}
\arguments{
\item{coefs}{cofficients from glm}

\item{newdata}{environmental data in}

\item{type}{of link defaults to "response"}

\item{varsin}{which parameters to include}
}
\value{
predictions
}
\description{
calculate predicitions from glm
}
