% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_functions.R
\name{get_chains}
\alias{get_chains}
\title{Import and bind chains for estimation}
\usage{
get_chains(path, burnin, thin)
}
\arguments{
\item{path}{file path to estimation directory}

\item{burnin}{burnin iterations to remove}

\item{thin}{factor to thin chain by}
}
\value{
mcmc_out - data frame of estimation
}
\description{
This functions takes a file path, finds and imports all
the .csv files in that directory and then tries to bind them.
If specified, it will also remove a burnin period and thin.
}
