% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{get_pop_data_3d}
\alias{get_pop_data_3d}
\title{function to import and get pop data in array}
\usage{
get_pop_data_3d(path, c_country, dat)
}
\arguments{
\item{path}{path to data}

\item{c_country}{list of countries}

\item{dat}{environmental and occurrence data}
}
\value{
pop1, pop3d, P_tot_2d, p_prop_3d
}
\description{
function to import and get pop data in array
}
