% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_mcmc_functions.R
\name{GLM_MCMC}
\alias{GLM_MCMC}
\title{mcmc for the GLM parameters}
\usage{
GLM_MCMC(Niter, name_dir, pars_ini, x, y, plot_chain)
}
\arguments{
\item{Niter}{number of interations}

\item{name_dir}{where to save output}

\item{pars_ini}{vector of inital parameters for GLM}

\item{x}{covariates}

\item{y}{explanatory variable}

\item{plot_chain}{whether to plot the first parameter's chain (TRUE or FALSE)}
}
\value{
saves chain in a folder of name paste0(name_dir,"/","GLM_chain",fileIndex,"_output.csv")
}
\description{
mcmc for the GLM parameters
}
