% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_mcmc_functions.R
\name{GLM_MCMC_step}
\alias{GLM_MCMC_step}
\title{mcmc step for the GLM parameters}
\usage{
GLM_MCMC_step(param, x, y, chain_cov, adapt, accCurrent)
}
\arguments{
\item{param}{vector of parameters for GLM}

\item{x}{covariates}

\item{y}{explanatory variable}

\item{chain_cov}{covariance of chain up to now}

\item{adapt}{whether to use the chain covariance to adapt the proposal distribution is 0 or 1}

\item{accCurrent}{current posterior probability}
}
\value{
new param, new posterior probability, whether step was accepted or not
}
\description{
mcmc step for the GLM parameters
}
