% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_mcmc_functions.R
\name{GLMlike}
\alias{GLMlike}
\title{likelihood for the GLM parameters}
\usage{
GLMlike(beta, x, y)
}
\arguments{
\item{beta}{coefficients}

\item{x}{covariates}

\item{y}{explanatory variable}
}
\value{
log likelihood
}
\description{
likelihood for the GLM parameters
}
