% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_mcmc_functions.R
\name{GLMprior}
\alias{GLMprior}
\title{Prior distributions for the GLM parameters}
\usage{
GLMprior(param)
}
\arguments{
\item{param}{vector of parameters for GLM}
}
\value{
vector of prior probabilities
}
\description{
Prior distributions for the GLM parameters
}
