% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_mcmc_functions.R
\name{GLMproposal}
\alias{GLMproposal}
\title{Proposal distribution for GLM}
\usage{
GLMproposal(param, chain_cov, adapt)
}
\arguments{
\item{param}{vector of parameters for GLM}

\item{chain_cov}{covariance of chain up to now}

\item{adapt}{whether to use the chain covariance to adapt the proposal distribution is 0 or 1}
}
\value{
param_prop. The proposed parameters.
}
\description{
Proposal distribution for GLM
}
