% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_functions.R
\name{hpd_out}
\alias{hpd_out}
\title{Internal function to get hpd}
\usage{
hpd_out(mcmc_out_chunk, log_scale = TRUE)
}
\arguments{
\item{mcmc_out_chunk}{section of mcmc_out}
}
\value{
hpd of chunk
}
\description{
Internal function to get hpd
}
