% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{import_pop_data_LS2014}
\alias{import_pop_data_LS2014}
\title{old popualtion data import function}
\usage{
import_pop_data_LS2014(path, c_country, adm)
}
\arguments{
\item{path}{path to data}

\item{c_country}{names of countries}

\item{adm}{admin level}
}
\value{
population by age and year
}
\description{
old popualtion data import function
}
