% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{launch_env_dat}
\alias{launch_env_dat}
\title{load environmental data from path}
\usage{
launch_env_dat(filepath, c34, dat_full = NA, delete_surv_AGO = TRUE)
}
\arguments{
\item{filepath}{file path to data csv}

\item{c34}{list of countries}

\item{dat_full}{data if already imported, defaults to NA}

\item{delete_surv_AGO}{remove AGO from surveillance quality, defaults to TRUE}
}
\value{
depi and dat, dependant variable index and data
}
\description{
load environmental data from path
}
