% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{Make_aggregate_pop_vc}
\alias{Make_aggregate_pop_vc}
\title{aggregating population and vaccination data for serological surveys}
\usage{
Make_aggregate_pop_vc(pop1, vc2d, sero_studies, adm1s)
}
\arguments{
\item{pop1}{population data}

\item{vc2d}{vaccine coverage in 2-d}

\item{sero_studies}{list of serological studies}

\item{adm1s}{list of admins for each survey}
}
\value{
pop_agg, vc_agg
}
\description{
aggregating population and vaccination data for serological surveys
}
