% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{Make_aggregate_pop_vc_3d}
\alias{Make_aggregate_pop_vc_3d}
\title{Aggregate the population and vaccination data into an array}
\usage{
Make_aggregate_pop_vc_3d(pop1, vc2d, sero_studies, adm1s)
}
\arguments{
\item{pop1}{population data}

\item{vc2d}{vaccine coverage in 2-d}

\item{sero_studies}{list of serological studies}

\item{adm1s}{list of admins for each survey}
}
\value{
pop_agg3d, vc_agg3d
}
\description{
Aggregate the population and vaccination data into an array
}
