% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{Make_pop3d_Ptot2d_Pprop3d}
\alias{Make_pop3d_Ptot2d_Pprop3d}
\title{function to reorganise population data}
\usage{
Make_pop3d_Ptot2d_Pprop3d(pop2d, adm)
}
\arguments{
\item{pop2d}{2-dimensional population data}

\item{admin}{level}
}
\value{
pop3d, P_tot_2d, p_prop_3d ie. population as an array,
the total population in a matrix and the population proportion in an array
}
\description{
function to reorganise population data
}
