% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_functions.R
\name{mean_fun}
\alias{mean_fun}
\title{Internal mean function}
\usage{
mean_fun(dat, idx)
}
\arguments{
\item{dat}{vector to be summarised}

\item{idx}{index}
}
\value{
mean
}
\description{
Internal mean function
}
