% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_functions.R
\name{plot_glm_map}
\alias{plot_glm_map}
\title{plot a map of glm output alongside map of data}
\usage{
plot_glm_map(shp0, shp1, c34, dat, Est_beta, x, colours)
}
\arguments{
\item{shp0}{shapefile for Africa countries}

\item{shp1}{shapefile for Africa admin 1 units}

\item{c34}{list of countries}

\item{dat}{occurrence and environmental data}

\item{Est_beta}{estimated coefficients}

\item{x}{covariates}

\item{colours}{what colours to use- length = 1000}
}
\description{
plot a map of glm output alongside map of data
}
