% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_functions.R
\name{plot_prior_post}
\alias{plot_prior_post}
\title{Plot the prior and posterior deinsities}
\usage{
plot_prior_post(mcmc_out, prior_col = "red", glm_or_sero)
}
\arguments{
\item{mcmc_out}{chain}

\item{prior_col}{colour of prior distributions, defaults to "red"}

\item{glm_or_sero}{whether plotting glm parameters or shared serology, either "GLM" or "SERO".}
}
\value{
plot of prior and posterior distributions
}
\description{
Function to plot the prior and posterior densities for parameters.
Note that the prior distributions are hard coded here.
}
