% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_functions.R
\name{plot_sero_predictions}
\alias{plot_sero_predictions}
\title{plot the seroprevalence predictions with data and credible intervals}
\usage{
plot_sero_predictions(seroout, hpd_sero, pop_agg3d, dim_year, dim_age,
  inc_v3d, pop_moments_agg, vc_agg3d)
}
\arguments{
\item{seroout}{processed serology data}

\item{hpd_sero}{high probability density region of parameter estimates for product space estimation of seroprevalence models}

\item{pop_agg3d}{population aggregated in array form}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{inc_v3d}{incidence of vaccination in array form}

\item{pop_moments_agg}{aggregated population moments}

\item{vc_agg3d}{vaccination coverage aggregated in array form}
}
\value{
figure of all serological surveys and predictions
}
\description{
plot the seroprevalence predictions with data and credible intervals
}
