% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_functions.R
\name{plot_transmission_intensity}
\alias{plot_transmission_intensity}
\title{calculate and plot the transmission intensity across the African endemic region}
\usage{
plot_transmission_intensity(x, ii, seroout, params, dat, t0_vac_africa,
  dim_year, dim_age, p_prop_3d, P_tot_2d, inc_v3d, pop1, vc2d, varsin_nc,
  polydeg, R0_lookup, model_type, shp1, shp0, Countries, colours)
}
\arguments{
\item{x}{glm covariates}

\item{ii}{indices of glm parameters}

\item{seroout}{processed serology data}

\item{params}{estimated parameters for both serology and GLM from either R0 or Foi model}

\item{dat}{environmental and occurrence data}

\item{t0_vac_africa}{time of first vaccination for admins in Africa endemic zone}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{p_prop_3d}{prpoprtion of population in array form}

\item{P_tot_2d}{total population in 2d form}

\item{pop1}{population data}

\item{vc2d}{vaccination coverages in 2d form}

\item{varsin_nc}{covariates in use}

\item{polydeg}{degree of polynomial, defaults to 5}

\item{R0_lookup}{table of R0 and infection numbers for each admin}

\item{model_type}{either "R0" or "Foi"}

\item{shp1}{shapefiles at admin 1}

\item{shp0}{shape files at country level}

\item{Countries}{list of countries}

\item{colours}{vector of length 100}

\item{inc_vc3d}{incidence of vaccination in array form}
}
\value{
map of transmission intensity
}
\description{
calculate and plot the transmission intensity across the African endemic region
}
