% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{process_serology}
\alias{process_serology}
\title{extract key information from serological data}
\usage{
process_serology(Serology)
}
\arguments{
\item{Serology}{dataframe of serological data}
}
\value{
survey_dat, adm1s, sero_studies, study_years, vc_factor, t0_vac, no_sero_studies
}
\description{
extract key information from serological data
}
