% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0_model_functions.R
\name{R0_calculate_seroprevalence}
\alias{R0_calculate_seroprevalence}
\title{function to calculate aggregated seroprevalence at survey in R0 model}
\usage{
R0_calculate_seroprevalence(index_survey, R0, foi_const = 0, t0_vac,
  dim_year, dim_age, study_years, p_at_survey_3d, P_tot_survey_2d,
  inc_v3d_agg, age_groups, pop_moments_agg, vcfac_arg = 1, vac_eff_arg)
}
\arguments{
\item{index_survey}{index of serological survey}

\item{R0}{value of R0 for survey}

\item{foi_const}{constant force of infection to add to R0 model, defaults to 0}

\item{t0_vac}{time of vaccination, vector}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{study_years}{years of sero surveys, vector}

\item{p_at_survey_3d}{population proportion at survey as an array}

\item{P_tot_survey_2d}{population total at surveys}

\item{inc_v3d_agg}{incidence of vacciantion at surveys}

\item{age_groups}{age groups in survey}

\item{pop_moments_agg}{aggregated population moments at surveys}

\item{vcfac_arg}{vaccine coverage factor, defaults to 1}

\item{vac_eff_arg}{vaccine efficacy}
}
\value{
seroprevalence at age groups
}
\description{
function to calculate aggregated seroprevalence at survey in R0 model
}
