% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0_model_functions.R
\name{R0_recurrence_seroprev}
\alias{R0_recurrence_seroprev}
\title{function to calculate seroprevalence by recurrence relation in R0 model,
forms the basis of R0_recurrence_seroprev_survey and R0_recurrence_seroprev_whole}
\usage{
R0_recurrence_seroprev(R0, foi, foi_const = 0, time_limit, dim_year,
  dim_age, p_at, P_tot, inc_v, vac_eff_arg)
}
\arguments{
\item{R0}{value of R0 for survey}

\item{foi}{foi before vaccination}

\item{foi_const}{constant force of infection to add to R0 model}

\item{time_limit}{time of vaccination}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{p_at}{population proportion}

\item{P_tot}{population total}

\item{inc_v}{incidence of vaccination}

\item{vac_eff_arg}{vaccine efficacy}
}
\value{
all statuses in R0 model
}
\description{
function to calculate seroprevalence by recurrence relation in R0 model,
forms the basis of R0_recurrence_seroprev_survey and R0_recurrence_seroprev_whole
}
