% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0_model_functions.R
\name{R0_recurrence_seroprev_survey}
\alias{R0_recurrence_seroprev_survey}
\title{function to calculate seroprevalence at survey by recurrence relation in R0 model}
\usage{
R0_recurrence_seroprev_survey(R0, foi_const, t0_vac, dim_year, dim_age,
  p_at_survey, P_tot_survey, inc_v3d, pop_moments, vac_eff_arg)
}
\arguments{
\item{R0}{value of R0 for survey}

\item{foi_const}{constant force of infection to add to R0 model}

\item{t0_vac}{time of vaccination, scalar}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{p_at_survey}{population proportion at one survey}

\item{P_tot_survey}{population total at one survey}

\item{inc_v3d}{incidence of vaccination at one survey (from inv_v3d_agg)}

\item{pop_moments}{aggregated population moments at one survey (from pop_moments_agg)}

\item{vac_eff_arg}{vaccine efficacy}
}
\value{
all statuses in R0 model for one survey
}
\description{
function to calculate seroprevalence at survey by recurrence relation in R0 model
}
