% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0_model_functions.R
\name{R0_recurrence_seroprev_whole}
\alias{R0_recurrence_seroprev_whole}
\title{function to calculate seroprevalence everywhere by recurrence relation in R0 model}
\usage{
R0_recurrence_seroprev_whole(adm, dat, R0, t0_vac_adm, dim_year, dim_age,
  p_prop_3d, P_tot_2d, inc_v3d, pop_moments_whole, vac_eff_arg)
}
\arguments{
\item{adm}{admin 1 unit of interest}

\item{dat}{environmental and occurrence data}

\item{R0}{value of R0 for survey}

\item{t0_vac_adm}{time of vaccination for admin, scalar}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{p_prop_3d}{population proportion}

\item{P_tot_2d}{population total}

\item{inc_v3d}{incidence of vaccination}

\item{pop_moments_whole}{aggregated population moments}

\item{vac_eff_arg}{vaccine efficacy}
}
\value{
all statuses in R0 model for admin
}
\description{
function to calculate seroprevalence everywhere by recurrence relation in R0 model
}
