% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERO_productspace_mcmc_functions.R
\name{Sero_Gibbs_MCMC}
\alias{Sero_Gibbs_MCMC}
\title{MCMC for Sero}
\usage{
Sero_Gibbs_MCMC(pars_ini, seroout, foi_const_surv, vc_agg3d, pop_agg3d,
  pop_moments_agg, dim_year, dim_age, p_at_survey, P_tot_survey,
  inc_v3d_agg, model_type, parameter_type, prob_Foi,
  posterior_distributions, ign, name_dir, Niter, run_id = 1)
}
\arguments{
\item{pars_ini}{vector of initial parameters for seroprevalence}

\item{seroout}{processed serological data}

\item{foi_const_surv}{vector of additional foi for each survey (R0 model)}

\item{vc_agg3d}{aggregated vaccination array}

\item{pop_agg3d}{aggregated population array}

\item{pop_moments_agg}{population moments aggregated}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{p_at_survey}{population proportion at time of serological survey}

\item{P_tot_survey}{total population at time of serological survey}

\item{inc_v3d_agg}{the incidence of vaccination (aggregated)}

\item{model_type}{either "R0" or "Foi"}

\item{parameter_type}{indexing for the different types of parameters}

\item{prob_Foi}{log of the model prior for Foi model}

\item{posterior_distributions}{distribution parameters for each model pseudo priors}

\item{ign}{sero surveys to ignore (good for bug hunting)}

\item{name_dir}{where to save the output}

\item{Niter}{number of mcmc iterations}

\item{run_id}{for parallel running- differentiates the different chains. Default = 1}
}
\value{
saves the output in name_dir
}
\description{
MCMC for Sero
}
