% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERO_productspace_mcmc_functions.R
\name{Sero_Gibbs_MCMC_parameter_step}
\alias{Sero_Gibbs_MCMC_parameter_step}
\title{MCMC parameter step for Sero}
\usage{
Sero_Gibbs_MCMC_parameter_step(param, seroout, foi_const_surv, vc_agg3d,
  pop_agg3d, pop_moments_agg, dim_year, dim_age, p_at_survey, P_tot_survey,
  inc_v3d_agg, chain_cov, adapt, model_type, parameter_type, accCurrent,
  prob_Foi, posterior_distributions, ign)
}
\arguments{
\item{param}{vector of parameters for seroprevalence}

\item{seroout}{processed serology data}

\item{foi_const_surv}{vector of additional foi for each survey (R0 model)}

\item{vc_agg3d}{aggregated vaccination array}

\item{pop_agg3d}{aggregated population array}

\item{pop_moments_agg}{population moments aggregated}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{p_at_survey}{population proportion at time of serological survey}

\item{P_tot_survey}{total population at time of serological survey}

\item{inc_v3d_agg}{the incidence of vaccination (aggregated)}

\item{chain_cov}{covariance of chain up to now}

\item{adapt}{whether to use the chain covariance to adapt the proposal distribution is 0 or 1}

\item{model_type}{either "R0" or "Foi"}

\item{parameter_type}{indexing for the different types of parameters}

\item{accCurrent}{current posterior probability}

\item{prob_Foi}{log of the model prior for Foi model}

\item{posterior_distributions}{distribution parameters for each model pseudo priors}

\item{ign}{sero surveys to ignore (good for bug hunting)}
}
\value{
next step in the metropolis-within-gibbs for the parameters (param, like, prior, accept)
}
\description{
MCMC parameter step for Sero
}
