% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERO_productspace_mcmc_functions.R
\name{SEROlike}
\alias{SEROlike}
\title{Likelihood distribution for Sero}
\usage{
SEROlike(param, seroout, foi_const_surv, vc_agg3d, pop_agg3d,
  pop_moments_agg, dim_year, dim_age, p_at_survey, P_tot_survey,
  inc_v3d_agg, model_type, ign)
}
\arguments{
\item{param}{vector of parameters for seroprevalence}

\item{seroout}{processed serology data}

\item{foi_const_surv}{vector of additional foi for each survey (R0 model)}

\item{vc_agg3d}{aggregated vaccination array}

\item{pop_agg3d}{aggregated population array}

\item{pop_moments_agg}{population moments aggregated}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{p_at_survey}{population proportion at time of serological survey}

\item{P_tot_survey}{total population at time of serological survey}

\item{inc_v3d_agg}{the incidence of vaccination (aggregated)}

\item{model_type}{either "R0" or "Foi"}

\item{ign}{indices of surveys to ignore in the likelihood- useful for bug hunting}
}
\value{
log likelihood
}
\description{
Likelihood distribution for Sero
}
