% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERO_productspace_mcmc_functions.R
\name{SEROlike_onesurvey}
\alias{SEROlike_onesurvey}
\title{likelihood for one serological survey in either model}
\usage{
SEROlike_onesurvey(model_type, transmission_intensity, seroout, pop,
  vc = 0, surveyIndex, foi_const = 0, dim_year, dim_age, p_at_survey,
  P_tot_survey, inc_v3d_agg, pop_moments_agg, vcfac_arg = 1, vac_eff_arg)
}
\arguments{
\item{model_type}{whether in R0 or Foi model}

\item{transmission_intensity}{value of foi or R0 for survey}

\item{seroout}{processed serology data}

\item{pop}{population in year and place of survey}

\item{vc}{vaccine coverage, defaults to 0}

\item{foi_const}{constant force of infection to add to R0 model}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{inc_v3d_agg}{incidence of vacciantion at survey}

\item{pop_moments_agg}{aggregated population moments at survey}

\item{vcfac_arg}{vaccine coverage factor}

\item{vac_eff_arg}{vaccine efficacy}

\item{age_groups1}{age groups in one survey}

\item{index_survey}{index of serological survey}

\item{p_at_survey_3d}{population proportion at survey as an array}

\item{P_tot_survey_2d}{population total at survey}
}
\value{
log likelihood for one serological survey from either model
}
\description{
likelihood for one serological survey in either model
}
