% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERO_productspace_mcmc_functions.R
\name{SEROprior}
\alias{SEROprior}
\title{Prior distribution for Sero}
\usage{
SEROprior(param, parameter_type, model_type)
}
\arguments{
\item{param}{vector of parameters for seroprevalence}

\item{parameter_type}{indexing for the different types of parameters}

\item{model_type}{either "R0" or "Foi"}
}
\value{
vector of prior probabilities
}
\description{
Prior distribution for Sero
}
