% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERO_productspace_mcmc_functions.R
\name{SEROproposal}
\alias{SEROproposal}
\title{Proposal distribution for Sero}
\usage{
SEROproposal(param, chain_cov, adapt, model_type)
}
\arguments{
\item{param}{vector of parameters for seroprevalence}

\item{chain_cov}{covariance of chain up to now}

\item{adapt}{whether to use the chain covariance to adapt the proposal distribution is 0 or 1}

\item{model_type}{either "R0" or "Foi"}
}
\value{
proposed parameters (param_prop) and proposed model type (model_type_prop)
}
\description{
Proposal distribution for Sero
}
