% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERO_productspace_mcmc_functions.R
\name{SEROpseudoprior}
\alias{SEROpseudoprior}
\title{Pseudo prior probability}
\usage{
SEROpseudoprior(model_type, param, posterior_distributions, prob_Foi)
}
\arguments{
\item{model_type}{either "R0" or "Foi"}

\item{param}{vector of parameters for seroprevalence}

\item{posterior_distributions}{distribution parameters for each model pseudo priors}

\item{prob_Foi}{log of the model prior for Foi model}
}
\value{
pseudo prior probabilities (pseudoprior) and model prior probability (modelprior)
}
\description{
Pseudo prior probability
}
