% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_functions.R
\name{setup_modelprior}
\alias{setup_modelprior}
\title{function estimate a value for the model prior in order to have even mixing.
Note, this is based on initial parameter values so may need adjusting.}
\usage{
setup_modelprior(pars_ini, seroout, foi_const_surv, vc_agg3d, pop_agg3d,
  pop_moments_agg, dim_year, dim_age, p_at_survey, P_tot_survey,
  inc_v3d_agg, parameter_type, ign)
}
\arguments{
\item{pars_ini}{vector of initial parameters for seroprevalence}

\item{seroout}{processed serology data}

\item{foi_const_surv}{vector of additional foi for each survey (R0 model)}

\item{vc_agg3d}{aggregated vaccination array}

\item{pop_agg3d}{aggregated population array}

\item{pop_moments_agg}{population moments aggregated}

\item{dim_year}{years of interest}

\item{dim_age}{ages of interest}

\item{p_at_survey}{population proportion at time of serological survey}

\item{P_tot_survey}{total population at time of serological survey}

\item{inc_v3d_agg}{the incidence of vaccination (aggregated)}

\item{parameter_type}{indexing for the different types of parameters}

\item{ign}{indices of surveys to ignore in the likelihood- useful for bug hunting}
}
\value{
Estimate of log model prior for Foi model
}
\description{
function estimate a value for the model prior in order to have even mixing.
Note, this is based on initial parameter values so may need adjusting.
}
