% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_functions.R
\name{setup_pseudoprior}
\alias{setup_pseudoprior}
\title{function estimate a value for the pseudo prior in order to have even mixing.
Note, this is based on initial parameter values so may need adjusting.}
\usage{
setup_pseudoprior(pars_ini, posterior_distributions, scale_model_type,
  tolerance = 0.01)
}
\arguments{
\item{pars_ini}{vector of initial parameters for seroprevalence}

\item{posterior_distributions}{distribution parameters for each model pseudo priors}

\item{scale_model_type}{the model whose posterior distribution sd is to be scaled}

\item{tolerance}{how close the difference between the pseudo prior probabilities 
shoud be at the initial parameters. Defaults to 0.01}
}
\value{
Estimate of scaling factor of scale_model_type posterior distributions sd- 
        assuming they are normally distributed
}
\description{
function estimate a value for the pseudo prior in order to have even mixing.
Note, this is based on initial parameter values so may need adjusting.
}
