% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup_functions.R
\name{transform_into_vc3d}
\alias{transform_into_vc3d}
\title{get vaccine data into an array}
\usage{
transform_into_vc3d(vc2d, adm)
}
\arguments{
\item{vc2d}{2-dimensional vaccination data}

\item{adm}{admin level}
}
\value{
vc3d
}
\description{
get vaccine data into an array
}
